requirements.txt
flake8==3.9.2
importlib-metadata==4.5.0
libusb==1.0.23b7
libusb1==1.9.2
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyQt6==6.1.1
PyQt6-Qt6==6.1.1
PyQt6-sip==13.1.0
typing-extensions==3.10.0.0
zipp==3.4.1
